allow programmatic json access from other api
