+++ src/app/resources/page.jsx
@@ -21,7 +21,7 @@ const ResourcesPage = () => {
 					<ImageLinkCard
 						title="Workshop Archive"
 						description="Check out our past workshops!"
-						alt="A man (Alexander Dean Cybulski) speaks at an event with a projector in the background."
+						alt=""
 						href="/resources/workshops"
 						height="auto"
 						image={workshopImage}
@@ -31,28 +31,19 @@ const ResourcesPage = () => {
 					<ImageLinkCard
 						title="Logo Download"
 						description="Download our bracket logo in crisp quality!"
-						alt="Google Developer Student Clubs UTM logo over a blue background with clouds"
+						alt=""
 						href="/resources/logo-download"
 						height="auto"
 						image={Brand}
 					/>
 				</Grid>
-				{/* <Grid item>
-						<ImageLinkCard
-							title="Tech tips"
-							description="Linus from Linus Tech Tips is a GDSC UTM member! :) Check out his tips here!"
-							alt="" // todo
-							href="/resources/tips"
-							image={BannerImg}
-						/>
-					</Grid> */}
 			</Grid>
 
 			{/* opting to keep FAQ in this page to keep it more visible */}
 			<h2 className="resources" id="faq">
 				Frequently Asked Questions
 			</h2>
-			<FAQList faq={faq} />
+			<FAQList {...faq} />
 		</HeroLayout>
 	);
 };
